gin: move gin_resources from base EP to RDMA EP
Move the gin_resources pointer and accessors from the
transport-generic nccl_net_ofi_ep_t to the RDMA-specific
nccl_net_ofi_rdma_ep_t, since GIN is only used with the
RDMA transport.

Remove the set_gin_resources(nullptr) call from
ep_holder destructor as it is unnecessary — the resources
object is about to be deleted.
