+++ include/nccl_ofi.h
@@ -752,20 +752,6 @@ public:
 		return *domain;
 	}
 
-	/**
-	 * Get pointer to the gin resources associated with this endpoint, or
-	 * nullptr if there are no associated GIN resources.
-	 */
-	inline nccl_ofi_gin_resources *get_gin_resources()
-	{
-		return gin_resources;
-	}
-
-	inline void set_gin_resources(nccl_ofi_gin_resources *gin_resources_arg)
-	{
-		this->gin_resources = gin_resources_arg;
-	}
-
 protected:
 	/**
 	 * @brief	Virtual destructor.
@@ -806,11 +792,6 @@ protected:
 	 * zero. sendrecv_release_ep() releases the resources if the
 	 * reference counter is decreased down to zero. */
 	int ref_cnt;
-
-	/**
-	 * Associated GIN resources object
-	 */
-	nccl_ofi_gin_resources *gin_resources = nullptr;
 };
 
 enum nccl_net_ofi_comm_type_t {
