+++ include/nccl_ofi_rdma.h
@@ -247,6 +247,7 @@ typedef struct nccl_net_ofi_flush_data {
 class nccl_net_ofi_rdma_device_t;
 class nccl_net_ofi_rdma_domain_t;
 class nccl_net_ofi_rdma_ep_t;
+class nccl_ofi_gin_resources;
 
 class nccl_net_ofi_rdma_device_rail_t;
 class nccl_net_ofi_rdma_domain_rail_t;
@@ -1088,6 +1089,20 @@ public:
 	}
 
 	/**
+	 * Get pointer to the gin resources associated with this endpoint, or
+	 * nullptr if there are no associated GIN resources.
+	 */
+	inline nccl_ofi_gin_resources *get_gin_resources()
+	{
+		return gin_resources;
+	}
+
+	inline void set_gin_resources(nccl_ofi_gin_resources *gin_resources_arg)
+	{
+		this->gin_resources = gin_resources_arg;
+	}
+
+	/**
 	 * @brief Return endpoint rail with index `rail_id`
 	 */
 	inline nccl_net_ofi_rdma_ep_rail_t *rdma_endpoint_get_rail(uint16_t rail_id)
@@ -1335,6 +1350,10 @@ protected:
 				nccl_net_ofi_rdma_cq_rail_t *cq_rail,
 				uint32_t tclass);
 
+	/**
+	 * Associated GIN resources object
+	 */
+	nccl_ofi_gin_resources *gin_resources = nullptr;
 };
 
 /*
